refactor: change file and variable names
