+++ src/controllers/index.js
@@ -1,2 +1,2 @@
 export * as usersController from './usersController.js';
-export * as postControllers from './postControllers.js';
+export * as postsController from './postsController.js';
