+++ src/controllers/postsController.js
@@ -1,8 +1,8 @@
-import { postRepository } from '../repositories/index.js';
+import { postsRepository } from '../repositories/index.js';
 
 async function timeline(req, res) {
   try {
-    const posts = await postRepository.getPost();
+    const posts = await postsRepository.getPost();
     res.status(200).send(posts);
   } catch (error) {
     res.sendStatus(500);
@@ -13,7 +13,7 @@ async function sendPost(req, res) {
   const { content, postLink } = req.body;
   const userId = res.locals.data.id;
   try {
-    await postRepository.savePostInDatabase(userId, content, postLink);
+    await postsRepository.savePostInDatabase(userId, content, postLink);
     res.sendStatus(201);
   } catch (error) {
     console.log(error);
